requirements.txt
numpy==1.26.4
matplotlib==3.10.0
scipy==1.14.1
dtaidistance==2.3.12
tensorflow==2.18.0
scikit-learn==1.6.0
torch==2.5.1
statsmodels==0.14.4
tslearn==0.6.3
seaborn==0.13.2
mgzip==0.2.1
protobuf==3.20.0
Jinja2==3.1.5
GitPython==3.1.44